requirements.txt
Django==2.1.7
django-cors-headers==2.4.0
django-extensions==2.1.5
django-rest-auth==0.9.3
djangorestframework==3.9.1
psycopg2==2.7.7
python-decouple==3.1
pytz==2018.9
six==1.12.0
